MBS-6518: Notice when wiki doc pages fail to be received and retry
The wiki docs server uses chunked transfer encoding, and occassionally
gets stuck when delivering some of these chunks. LWP::UserAgent does
respect its timeout, however it retains the initial status code and deems
the request as successful, even if the entire body wasn't received.

We now check the 'X-Died' header, which is set when the response body
is partial, and if it exists and is a 'read timeout', we retry the request.
Requests are attempt as many times as they will fit within a 25 second window.