MBS-2301: Preserve disc IDs in release editor when using a duplicate as a starting point
If the user decides to use another release as a starting point and
they have also seeded the release editor with a TOC, we now ensure
the TOC is retained for the first medium, and than the first medium
has exactly the same amount of tracks as the TOC suggests. If the
release to use as a starting point doesn't have enough tracks on
medium one, then extra tracks are added. If it has too many tracks,
then these extra tracks are discarded.