MBS-6703: Correctly display cover art types in remove cover art edits
If the cover art hadn't actually been deleted, then the $artwork variable
is set from $loaded - because we can still load the artwork. However,
loaded artwork doesn't have any types set, so the types array remains
empty. It should be possible to use => ['CoverArtType'] to augment this
loading, but I can't get that to work.

As we're loading CoverArtTypes by hand anyway to display data after
the edit passes, I decided to simply always use that data and set this
attribute manually.