MBS-6748: Do not crash when showing relationship edits where attributes have since been deleted
We have a problem in that edits currently store attribute IDs, and
try and resolve these into attribute objects when the edit is displayed.
It's also assumed that these attribute IDs do resolve - but of course
attributes are occasionally removed.

Now we filter out only successfully loaded attributes, and if this
filtering discarded some attribute IDs, we display a message that
indicates that the edit isn't quite representative of what happened.