MBS-5647: Correct release header display in release editor
The template for the duplicates tab assigns to a variable called
release. For some (inane) reason, Template Toolkit doesn't have
lexical scoping - so this is a global variable. Furthermore, the
content of a wrapper is evaluated *before* the wrapper. Hence the
end result is that when the wrapper is ran, the release variable
is set, so the layout wrapper assumes the user is editing a release.

This commit uses a different variable name, and hopes that no one
ever needs to use this variable name again.