MBS-6067: Always work with an inc object
If the user browses /artist/mbid?query=foo, it should really be an
incorrect request - because you can't query an MBID. However, making
that so is unfortunately a lot of work. In the meantime, I am
addressing a current exception that gets through.

Because ?query= does match a valid branch of inc arguments, there
is no failure when parsing the request - it just means nothing
ever assigns to $inc. However, the OAuth checks assume that $inc
is a hash reference, and autovivification turns $inc into an empty
hash reference.

Later, various other parts of the web service assume that $inc exists
and is an object, not a hash reference - and thus there is an exception.
I've instead opted to have $inc default to be an empty inc object.