MBS-3841: Move JSON release generation to a [%- PERL -%] block
Unfortunately, Template Toolkit doesn't let you pass undef to a
filter function - instead it gets silently coerced to the empty
string. But the empty string is different to 'null' in JSON, so
we end up getting strange behaviour when displaying null track
lengths.

By using a [%- PERL -%] block, we get sligthly better tools for
abstraction (so the generation is shorter), but more importantly -
we can preserve undef.