Use bool_or in merge_boolean_attributes
The previous behavior would set the column to 'TRUE' if any of the
entities being merged had 'TRUE' in that column. However, if all
rows were 'FALSE', then no rows will be considered for the merge
and hence the function would attempt to set the new value to 'NULL'.

Instead, use bool_or() aggregation, which is both more concise, but
more importantly - has the correct behavior in all cases.