MBS-3820: Factor out EditNote stuff to Role::EditNote; use in release editor
The parent commit introduced validation on the 'edit_note' field,
allowing it to be configurable whether or not it is required. The
release editor also has conditional validation on the edit note field,
but defines that itself. This commit unifies the two by creating
Form::Role::EditNote, and having both Form::Role::Edit and
Form::ReleaseEditor::EditNote compose this role.