MBS-6824: Coordinate field is now a single lat/long field
This field takes a latitude and longitude as a single string, and
parses it appropriately. Currently supported are formats in degrees
(only), and degrees with minutes and seconds. Direction indicates
N, S, E, W are are supported. If the field does not parse, then the
field is flagged as an error.