MBS-6785: Use the tagger iframe if window.opera exists
Opera doesn't seem to allow loading localhost as an image with its
security policy, so the fix for mixed-content introduced in
commit 2fef85c causes the tagger button to have no effect on Opera.

This commit conditionally branches on the 'window.opera' object,
which (should) only be set if the browser is Opera. If the browser
is Opera, we use an approach similar to the old iframe approach,
though we do so dynamically. If window.opera does not exist, we
continue to use the new Image technique.