MBS-6902: Filter object_to_ids to only operate on defined objects
The code makes the assumption that the objects will be defined, so
making this explicit in the iterator avoids an ISE if this method
gets passed an undefined value.

In the case of MBS-6902, the undefined values were coming from
Data::Edit->load_all's 'post loading' functionality, which assumed
that all objects could be successfully loaded in order to later
augment the objects with extra information. It's not guaranteed
that all these objects will load, so I'm now filtering out those
that are undefined.