MBS-6916: beta: No dropdown in inline search when there are no results
Adding the action buttons in _renderMenu was problematic because that method
isn't even called if there are no results (so it never gets a chance to add
the "(No Results)" button). So the buttons are now added in _lookupSuccess.

I also made a small adjustment to _searchTimeout so that the menu is properly
closed when the search input is cleared.