MBS-6918: artist credit changes break /release/add on beta
Problem 1:

I placed a copy of submitted_release_artist in the stash, but I added the
relevant line to prepare_tracklist in *ReleaseEditor::Edit*, which obviously
won't affect ReleaseEditor::Add. Now the line is in the main prepare_tracklist
function. (I haven't found out how this was working locally before, because I
specifically tested /release/add, but I must've changed something at the last
minute.)

Problem 2:

MB.Control.ArtistCredit expected initialData to not be an empty list. I'm not
sure if it's possible to be now that I've fixed problem 1, but just to be on
the safe side, I've modified Form::Field::ArtistCredit and
MB.Control.ArtistCredit to not return an empty list and not assume it's
dealing with a non-empty list, respectively.

Problem 3:

This escaped my initial testing. If you parse a tracklist with no artists,
the tracklist would no longer default to the release artist. That's now fixed.