MBS-6947: Load recording name in release JSON
When we create MB.entity.Recording objects for all the recordings
in a tracklist, it has the side effect of entering a new cache entry
keyed off the recording's MBID. I previously had the recording
information for stuff in the tracklist minimal - and didn't include
the recording's name. This meant that we would cache a recording
that had the empty string as its name.

Later, when we went to display recording relationships, if we
referred to the recording that is the same as the recording of
a track in the tracklist, we would get back an object with no name.

To fix this, I now load the recording name for all recording's in
the tracklist.