MBS-6986: Match recordings by MBID not working any more
After an MBID lookup, use options.select instead of currentSelection
for setting the selected entity. The former calls currentSelection
and also emits the lookup-performed event, which the release editor
listens on.

This also removes code that disables the input field while the entity
is being looked up, since it causes an unnecessary blur event and doesn't
serve any purpose (any current requests should be cancelled if the user
edits the field).