MBS-6978: can't close inline search
Instead of canceling blur events to prevent the menu closing after a click
event (which causes problems where the menu can't *ever* close), this switches
to a different approach.

If the user selects an action with the enter key, the behavior is the same as
before (preventDefault is called). But if the action is clicked, we catch
the click event early to perform the action, and call stopPropagation so that
the `select` method is never reached. Consequently, no blur event occurs.

Tested in Chromium 30, Firefox 26, and Opera 12.16 in the release and
relationship editors.