MBS-6983: Correct display of the 'video' column for merge recording edits
There are two paths for rendering a list of recordings, depending on
whether or not the recording came from the database or edit data. If the
edit came from the database the display was correct, but if the data
came from an edit then the alternative code path was executed which
ignored the `show_video` variable. This commit makes both code paths
consider the state of this variable.