MBS-6994: Hash work attributes to a canonical form
We hash work attributes by converting them to a JSON string and then
comparing these strings for equality. However, we didn't make sure that
the JSON itself was normalized into a canonical form. Specifically,
sometimes the type_id was a string, and sometimes it was an integer -
thus the comparison fails. This would result in two work attributes
being created, rather than work attributes being edited (or left unchanged).