MBS-6952: Correctly render the 'various artists' warning
The various artists warning is shown when the "Various Artists" artist
appears in a tracklisting. However, the actual route to making this
decision is quite complicated. At the top level, this decision is made
on a per-track basis, by the track subscribing to its own artist credit
and watching it change. It seems a track's artist credit is set three
times on the tracklist page:

1. Defaulted to the release artist
2. Cleared
3. Set to the actual track artist

In the release reported in MBS-6952, the release artist is "Various
Artists", so at step 1 the warning banner is shown. However, at step 3
track artists are *not* "Various Artists" so the banner should be
hidden. However, the call to update the banner only happens if the track
artist is "Various Artists". Thus the banner can only be turned on and
never off.

Instead of conditionally updating the banner, I have removed the entry
condition - it seems that the banner update itself checks whether or not
to show itself. The logic to toggle the banner is now called from a
subscription to the track artist credits (directly, rather than
implicitly as it was before).