Don't require missing AC fields for existing entities
If a form is editing an existing entity and the AC field is entirely
missing (as opposed to existing but being empty), the field will *not*
be required. The form will see this is as no changes being made. This
behavior allows bots and browser scripts to function properly (i.e.
environments where AC fields aren't generated).