MBS-7024: Truncate updates in replication packet 72901
replication-72901 contains an insert/update to release_group with a
release group name that is 12k characters long. However, we also have a
functional index on musicbrainz_collate(name), and for versions of
libicu at least >= 4.8, the result of musicbrainz_collate(name) is
larger than what can be stored in PostgreSQL indexes (older versions of
libicu produce a smaller result and can load this replication packet).

To work around this problem, the offending large insert/update column
has been hard-coded to be truncated to 1000 characters, which allows
replication to continue.