MBS-6928: Allow adding tracklists with VA as a release artist
When a VA release is added through the "submit CDTOC" function, the
tracks are not assigned a track artist. This leads to a JavaScript error
when the tracklist page is submitted, as 'artistChanges's is called, but
there is no 'from' artist, thus an exception is thrown. Now, if there is
no from artist, we simply return 'true' - there are changes to the
track's artists.