MBS-6772: Fix cover art reordering
The event handlers for the buttons to move images forward/backward on
the "reorder cover art" page stored a reference to the div.editimage
element containing the buttons. This reference, however, was
inadvertently shared between all "forward" buttons and all "backward"
buttons, respectively (two copies in total). The first image moved in
either direction won the race, and afterwards all buttons (for that
direction) of all images operated on that first image instead of their
corresponding image.

This is fixed by introducing a local (unshared) variable in the
handler function.