MBS-7077: beta: sizing issues with dialogs/autocompletes in relationship editor
It seems there's a bug with jQuery UI where if you have both "width: auto"
and "maxWidth" set, it won't respect maxWidth. I'll have to update our jQuery
UI and see if this is fixed. In any case, adding a max-width rule to the CSS
works and makes the dialog not take up the whole screen if there's really
long text.

The "max-width: 50%" we had for the autocomplete menu wasn't ideal for menus
inside dialogs (because it was intended to be 50% of the page, but results
in 50% of the dialog). So I changed the max-width to 500px instead.

Also removed a z-index rule that was a hack that only applied to the previous
dialog code.