MBS-7084: Update to_merge on /:entity/merge pages
The work on MBS-7057 meant that entities were only loaded from the
session once to avoid some odd state changed bugs. This meant that the
extra information loaded on the /:entity/merge confirmation page
was loaded, but never used in the actual rendered page as that
referenced a different variable.

This commit keeps most of the code the same, but on the merge confirmation
page we make sure to reset $c->stash->{to_merge} with the new set
of entities.