MBS-6804: Don't make assumptions about iTunes names
iTunes URLs may have a part with the item name in a human-readable
form. This part is removed by the URL cleanup code, which makes
assumptions about the characters allowable in this part. These
assumptions have repeatedly turned out to be wrong (MBS-6695, '+';
MBS-6743, '*'; MBS-6804, '$').

We could yet again add the character that was found missing this time
and risk that the list is still incomplete. On the other hand, we can
simply allow any character and still match the trailing "/id012345"
without ambiguity. The latter seems more future-proof, so this is what
is done here.