MBS-7061: Prevent network providers/proxies from stripping HTML comments
Knockout.js uses "virtual elements" (which are just HTML comments) very
heavily, so removing the comments breaks a lot of the website.

This adds a Cache-Control: no-transform header to page responses.
http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.5

While I'm not sure of a means to test this myself, I added the fix to my
sandbox and had the author of MBS-7061 load up a release page on my sandbox
and someone else's without the no-transform header (Ian's), and they reported
my sandbox rendered the page correctly while Ian's did not.