Reorder cover art: calculate positions differently
On the Reorder Cover Art form, set the values of the hidden input
fields giving the new position in one pass immediately before
submitting.

Previously the values were updated whenever an image was moved, by
swapping the values on the moved image and its neighbour. This was
less clear (in my opinion) and could only cope with simple movements
(one place to the left/right).