Add cover art: determine position differently
On the Add Cover Art page, determine the position where the new images
are to be inserted only when the form is submitted. This has the same
advantages as the corresponding Reorder Cover Art change in the
previous commit.

A small helper function set_position is used because the same code is
needed twice, for the AJAX uploader and the old one.