MBS-7139: Fix table sorter for statistics pages
The countries and languages tables on the statistics pages use the
jQuery tablesorter plugin so that the user can choose by which columns
the list should be sorted. However, this was broken in commit
5455da396170824bb49cbeb10655fb5a11dc45f3 that removed several
unnecessary document-ready handlers; because Tablesorter.js continued
to use such a handler, it ran after other code that depended on it and
failed.

This is fixed by removing the document-ready handler also in
Tablesorter.js. Additionally, MB.Control.Tablesorter is changed to
accept parameters such that it can be applied differently to separate
tables that share one page.

The commit is best viewed ignoring whitespace changes.