MBS-6265: Fix option to language table sorter
The sortList option to the table sorter for the language statistics
table was [[4,2,3,1]], which apparently was meant to select a default
sort order of column 4, then column 2, etc. This is not what the
tablesorter plugin expects, though; instead, the option should contain
a list of pairs [column, isDescending].

Fix this and also simplify the sort order (first descending by total
number, then by name). Also use this new sort order for the countries
table.