MBS-6959: Fix string test in link_entity macro
The link_entity macro displays a possible disambiguation comment
after the generated entity link if the entity name is not replaced by
a different link text. This was implemented with an "UNLESS text"
directive. Since the string "0" is considered false by Template
Toolkit, it could cause the disambiguation to be added even though a
replacement text had been present.

In particular, this caused a spurious disambiguation comment to be
displayed in the countries statistics table, where the former country
of Serbia and Montenegro has both a disambiguation comment and
"0" artists.

An explicit string comparison "IF text == ''" is used instead to fix
this.