MBS-7154: Attach CD TOC : some"show tracklist" buttons don't work
The script for toggling the tracklists gets included inside a FOR loop,
so many click handlers get registered for each button. Upon clicking
a button, all of the handlers registered after that specficic row
would get called, having a weird domino effect where the tracklist
would toggle back and forth a huge number of times until it stopped
on "show" or "hide" depending on being odd or even in the list.

This introduces MACROs for the release rows and scripts so that the
templates can include these separately in the markup.

This also fixes a few colspans that cause the barcode column to not
fit into the table.