MBS-7190: Prefer target when merging release events
Upon a release merge, when there are conflicting dates for the release
event in one country, the version in the merge target should be
preferred. The way the relevant ORDER BY clause was constructed
accomplished the opposite instead: A Boolean false value is
interpreted as 0, which sorts before 1 (true) in a default ascending
sort.

The sort order is changed to descending in order for true values to
appear first.