Add allow_auto_edit to Edit::Medium::Create
Problem: Add-medium edits should be auto-edits for new releases. The
current release editor has no problem here, because it creates
everything in one request/submission. But if we're creating each edit
separately/asynchronously, we need a way for the edit to still be
applied automatically.

Solution: In allow_auto_edit, return 1 if the release we're adding the
medium to was itself added by the same editor less than an hour ago
(and it's still open).