MBS-7100: Options and links for VA/all/SARs were very confused
I've gone through and listed the 32 possibilities and what each one
should show. Hopefully, in general, this should lead to coherent links,
though it does lead to quite a monster of a nested if statement. Filters
are taken to always mean "don't change my VA/all releases preference".
Recordings are searched for only if there is no filter and VA is not
requested. Otherwise, respecting preferences for VA/all, it falls back
first to all releases, and then to VA releases, and then to all VA releases.