MBS-7232: Beta: Relationship editor doesn't confirm leaving
window.onbeforeunload was always being set to undefined in the
"changes" subscription, which should've been removed in a previous
commit but was left in.

This commit also simplifies some code paths, standardizes on some
property names (s/media/mediums), and exposes the releaseLoaded
method so that tests can use it directly.