Fix data-loss in 2c610293, add to Edit::Release::Edit
In the _edit_hash functions, falling back to '' for undefined comments
happens even if a comment doesn't exist in the edit data. So, if just
an entity name was changed, the comment would be silently cleared
without showing up in the edit diff.

Expand the general behavior here to Edit::Release::Edit, since it was
forgotten in 2c610293.

In the JavaScript, no longer encode empty comments as null. Use empty
strings instead. This avoids all of the above entirely.