Better track parser matching to existing tracks
The current code will loop through all the parsed track data, and then
loop through all the existing tracks to find a match to whatever was
parsed. If the track it matches with was already used, however, it
doesn't check if the match is better than the previous one. The new
code does. Additionally, we now first check if there's a good match to
the current track in the same position. If there is, we use that and
don't perform any further matching.

Some functions and constants related to metadata matching were moved
around as part of this patch.