Revert "Update the track parser text as the user changes the options"
This reverts commit c0fb18a773a2b3cb6569abf690ac3294b221224b.

After discussion, it seems this is more annoying than useful. If you
forget to check an option (say, for track numbers) before parsing,
it'll add those numbers to the track titles, but then if you re-open
the parser dialog and check the option, it'll add duplicate numbers to
the tracks.