MBS-7285: Can't seed external links to entity editing pages other than the release editor
This allows adding things like edit-artist.url.0.text=http://www.example.com/&
edit-artist.url.0.link_type_id=1 to the query params for an entity editing
page.

If this is done for an existing entity, the relationships in the query params
are appended to the existing ones and do not replace everything.

If certain fields are seeded to an existing entity, such as relationship_id or
removed, then the behavior is undefined. It's not smart enough to "merge" the
existing and new data. But I'm guessing this might be okay for now as the
primary use will be to submit new URLs.