MBS-7322: Process both external link and ISWC edits
The EditExternalLinks role uses the post_creation hook. If this hook
already is in use, its old content is stored and called from within
the new hook function. The intention is to call both and return a true
value if either post_creation function made changes. However, the
short-circuiting feature of logical operators causes the external
links post_creation function to be called only if the original
post_creation function didn't make any changes.

Instead call edit_external_links unconditionally, then in a separate
step combine the return value with that of the original hook function.