Use EditExternalLinks in entity controllers directly
The fix for MBS-7322 in the previous commit uncovered a latent bug
where edits to remove external links were entered twice. This was due
to the EditExternalLinks role being included twice per entity
controller, namely via the Create and Edit roles; each time, the
"around" method modifier for edit_action was activated.

As proposed by bitmap on IRC, the EditExternalLinks role is removed
from the Create and Edit roles and instead included directly in the
controller for each core entity (except Release, which is handled by
the special Release Editor, and URL). The has_url_relationships
parameter is rendered unnecessary and therefore removed.