ISWC code not shown on ISWC page
Exactly the same as the previous commit, just for ISWC instead of
ISRC: Commit eca0daa3010a6b1ffa21b4ef1b082ef919782e24 changed the
link_iswc macro to only operate on an ISWC object, not on a mere
string. However, the template for ISWC pages was not fixed
accordingly.