MBS-7328: Improve the recording search in the release editor
The queries it sends no longer require the recording artists and
lengths to all be matched. Before, the logic was:

recording:foo AND arid:gids AND dur:[1 to 1000]

Now the logic is:

((recording:foo AND arid:gids)^2 OR recording:foo) AND
    dur:([1 TO 1000] OR \-)

The ^2 is supposed to boost the first subquery so that if the title AND
artist ids are matched, that's preferred over just the title. That
seems to function okay.

The escaped "-" in the dur field is to specifically include recordings
without any length. Thanks to Ian for figuring out how to do that for
me.