MBS-7356: Include a timestamp on the ISE page
When an Internal Server Error occurs, the information from the ISE
page is often pasted into an issue tracker ticket, to help with
debugging. It can be useful to know when exactly the ISE happened in
order to compare with the open/close times of any relevant edits, so
include such a timestamp in the ISE page.