MBS-7373: Edit diff for reordering is misleading
This changes the logic used to determine track MBID/recording changes.

Since you can't "reuse" a track from another release (at least not
through our UI), the only way you'll get a new track MBID right now
is if you actually remove a track and add a new one. So the logic
is now to simply check if the new track data for a certain position
already has an id (if it does, it was moved, so it's not changing).

To detect recording changes, it now does a set difference between
all the previous recordings and all the new recordings for existing
tracks. So the recordings that changed for existing tracks are
$new - $old.