MBS-7372: Release Editor: CD TOC x Track Parser breaks horribly
If you're adding a release from a TOC and you parse less tracks than
the TOC has, it pads the tracklist with empty tracks. Except it
doesn't, because it uses {} as an empty track instead of an actual
track instance. So later when other code tried to call track methods
that don't exist on the object, things exploded.