MBS-7378: Inline search returns duplicate results when searching for the same text twice
The `term` variable stores the last text that was used to do a search. So
if you type something and then quickly revert it, it'll still do a search
because it compares `term` to the current input value even if the search
for the current `term` was canceled.

Now we keep a `completedTerm` variable that stores each term until a
search actually completes for it, and compare against that as well.

Additionally, we now run _.str.clean on the search terms/values so that
typing a space doesn't initiate a search until you add more text.

Finally, if another search is in progress, we only clear the timeout
if we're starting a new one (or an MBID was pasted). I *think* this may
solve some of the randomly-canceled search issues Nikki was noticing.