MBS-7380: Fix work attribute statistics
Apparently, the template for work attribute statistics was copied from
the work type section and not adjusted completely. Therefore, the line
for works without attributes displayed the corresponding figures for
works without a type, and the percentage for each work attribute
showed the figure for the last work type instead. This is changed to
use the correct, work-attribute, values.