Fix logic in 647535f
Changes are now replaced with insertions/deletions if the track ids
are different.

Instead of keeping track of changed recordings using a set difference
(which fails to find recordings that were switched around), we now
keep a map of ( track id => original recording id ), and see if any
of the new recording ids differ.