MBS-7367: beta: error when adding disc ID with new release
Another silly bug caused by 522a468.

tmpPosition is the position of a medium as it's being submitted
before any reorder edits, to avoid position conflicts with already-
existing mediums. The property only exists in those circumstances,
and is otherwise undefined. So now we fallback to position() in the
cases where tmpPosition is undefined.