MBS-5897: Give access to raw edit data
Introduce a new /edit/:id/data endpoint that displays the raw data for
the given edit, for cases where the regular edit display fails or is
not sufficiently detailed for debugging purposes.

The link from the regular edit page is intentionally inconspicuous
since it is targetted at developers and near-developers, not at normal
users.