Display more information about edits on ISE page
When an internal server error occurs on a page that is supposed to
show edits, give more information about these edits on the ISE page.
In particular, list the edits by ID with a link both to their regular
page and to the raw edit data, and say whether the display data was
successfully loaded or not (which may give a hint as to which edit has
caused the error).

To achieve this, the hash containing the edits is stashed as soon as
possible in the relevant controllers, before loading any additional
data, so that it is accessible for the ISE template.