MBS-7412: Beta: reorder medium edits are generated for new releases
The relevant change is in root/static/scripts/release-editor/fields.js
The rest is just to allow for testing.

`mediums.original` is supposed to store the medium data as it exists in
the database. For new releases, it should be empty. Otherwise the
avoid-position-conflicts-at-all-costs code will incorrectly use this
data.