Add a hook for filtering possible duplicates
Add a filter_duplicates hook to the CheckDuplicates role that can be
overridden by consumers and allows them to preprocess the list of
possible duplicates. This can be used to load additional data for
them, to remove entries from the list, and to order entries, among
others. The function may also decide whether a disambiguation comment
is required or not; this defaults to yes in order to retain the
previous behaviour.