MBS-7411: Require place disambiguation only if necessary
Only require the user to enter a disambiguation comment for a place if
its area is not enough for disambiguation, i.e. if another place with
a similar place is in the same area or if no area is given.

In addition, possible duplicates are sorted, with the most unlikely
duplicates (e.g. in a different country) coming last.