Allow models to be chained in edit foreign_keys
Extend load_all to support a list of space-separated models to be
chain-loaded. E.g., an edit can now request in its foreign_keys a
post-load of "Area AreaContainment" for a place, meaning that the area
is to be loaded, then the containment not for the original place, but
for the area loaded in the first step.