Add infrastructure for auto-including models
Extend load_all with a new default_includes function that modifies the
data structure holding the objects to be loaded, automatically adding
post-load models according to a list of "if we load this model, we
also want that model" associations.

This commit only provides the infrastructure for that. The list
remains empty.