Show areas with containing areas in edit display
When displaying an edit, make sure that places and areas are shown
with their containing areas. This makes use of the infrastructure for
automatically including post-load models that was added in the
previous commit. In addition, a lot of templates are changed to use
descriptive_link instead of link_entity.