MBS-7437: Fix removing an area in place edits
The form for place edits did not have all required infrastructure for
the area autocomplete field; in particular, the gid value was not set
when the form loaded. When the user cleared the text field in order to
remove the area, the hidden gid field was set to the empty string
accordingly, and it was expected that the relevant change event would
fire; which did not happen, because the field had been empty to begin
with. The area_id field therefore retained its value, meaning that the
removal was ignored when the form was submitted.