MBS-5552: Mark edits that merge URL entities
An "edit URL" edit that changes the URL to one that already exists in
the database is actually a merge, which may be hard to revert if wrong
and therefore deserves extra scrutiny.

Two new fields are added to the edit data: the number of relationships
using this URL entity, and whether the edit is a merge. This
information is then presented in the edit view unless it is the
standard case (one relationship, no merge).