MBS-7450: Show coordinates for equatorial places
Places with at least one zero-valued coordinate had their coordinates
not displayed because the template checked for latitude and longitude
being true instead of the coordinates being defined. Similarly, the
"add place" edit class evaluated the latitude and would have stopped
showing coordinates on the equator as soon as MBS-7438 is fixed. (It
worked because "0.000000", unlike "0", evaluates to true.)