MBS-6508: Show removed areas in "edit" edits
In the edit display for artist, label, and place "edit" edits, areas
were not displayed if they had since been removed or merged into
another area. This made it impossible to distinguish if an edit deals
with such areas or with unset area fields. This is fixed by creating
an empty Area entity if an area ID is present in the edit data but the
corresponding area can't be loaded. Such an empty entity will be
displayed as "[removed]".