MBS-7455: Fix DeleteReleaseLabel ISE for /release/add previews
This occurs because after release.labels.original() is modified,
previews for label edits are regenerated. Since by then the release is
added but we don't have any ids for the new release labels, it sees
things as remove edits and generates previews accordingly.

One half-solution would be to simply stop previews while the release is
submitting, which is probably a good idea anyway, but doesn't help if
one of the submissions fails. So I changed /ws/js/edit to return the
release label ids that were created, and merge them in on the client
before resetting the original data.

This might possibly also fix weird issues where label edits were
submitted twice.