Fix display data building in "add artist" edit
The "add artist" build_display_data method was supposed to return an
empty string for sortname and disambiguation comment if these were
undefined, but the test for definedness was broken and always true.

This allows fixing the associated template so that it displays an
existing disambiguation comment even if the string evaluates to false
(such as "0").