MBS-7376: enter on <select> in Firefox doesn't behave right
Added a workaround for Firefox's weird <select> behavior again, similar
to the one for the relationship editor in Dialog.js, though this one's
shorter.

Selecting an <option> using the arrow keys requires you to press enter
to actually select it in Firefox (even though the menu isn't open). But
if the tab is switched immediately, the change doesn't seem to happen,
possibly because the <select> is hidden. So this adds a slight delay
which fixes the behavior for me (tested in Firefox 29).

Also tested in Chrome and Opera 12.10 to make sure they still work as
expected, even though they weren't affected by this bug.