Display entities related to an edit
On the /edit/:id/data page introduced by the previous commit, list the
entities that have this edit in their edit history (edit_artist etc.
tables). This is particularly useful when entities have been merged
after the edit was entered, because the entity IDs referred to in the
edit data are not updated but the edit history is.