Make the bubbles redraw less often in the release editor
The with binding causes all the HTML underneath it to redraw, as
explained by
http://www.knockmeout.net/2012/03/knockoutjs-performance-gotcha-1ifwith.html

This is bad because it makes to harder for userscripts to modify the
HTML. Anything added to the HTML is lost when a redraw occurs, because
knockout caches the original HTML for reuse.

I took the advice of the above article and made sure to "use the with
binding in smaller, targeted areas" where possible.

Most notably, the button sets in the AC editors no longer redraw.